add refactor to filters
